Refactor program and use dict for tasks
After reading through Clean Code, I realized a lot of my functions were
just way too big. I went through and tried to reduce functions to their
most declarative, enuring their abstraction is all on the same level,
and creating small stubs and subfunctions. The goal of this was to
reduce functions into doing one thing and one thing only.

I also changed how tasks are represented. They were previously
represented as tuples, which I had to remember what index was what and
destructure it all the time. Putting the task in a dictionary with all
of the values named was a much more readable and maintainable solution
to that problem.